Refactor exportCompressedRegistryImage
