+++ backend/src/services/mirroredModel.ts
@@ -1,4 +1,5 @@
 import { createHash } from 'node:crypto'
+import zlib from 'node:zlib'
 
 import archiver from 'archiver'
 import * as fflate from 'fflate'
@@ -39,6 +40,7 @@ import {
   setLatestImportedModelCard,
   validateMirroredModel,
 } from './model.js'
+import { getImageBlob, listImageTagLayers } from './registry.js'
 import { getAllFileIds, getReleasesForExport, isReleaseDoc, saveImportedRelease } from './release.js'
 
 export async function exportModel(
@@ -97,6 +99,78 @@ export async function exportModel(
   }
   zip.finalize()
   log.debug({ modelId, semvers }, 'Successfully finalized zip file.')
+}
+
+export async function exportCompressedRegistryImage(
+  user: UserInterface,
+  modelId: string,
+  imageName: string,
+  imageTag: string,
+  logData: Record<string, unknown>,
+  metadata?: ExportMetadata,
+) {
+  // get which layers exist for the model
+  const layers = await listImageTagLayers(user, modelId, imageName, imageTag)
+  log.debug('Got image tag manifest', {
+    modelId,
+    imageName,
+    imageTag,
+    layersLength: layers.length,
+    layers: layers.map((layer: { [x: string]: any }) => {
+      return { size: layer['size'], digest: layer['digest'] }
+    }),
+    ...logData,
+  })
+
+  // setup gzip
+  const gzipStream = zlib.createGzip({ chunkSize: 16 * 1024 * 1024, level: zlib.constants.Z_BEST_SPEED })
+  // default `['close', 'error', 'prefinish', 'finish', 'end', 'readable']` plus one `'error'` emitter per layer
+  gzipStream.setMaxListeners(6 + layers.length)
+
+  // start uploading the gzip stream to S3
+  const path = `beta/registry/${modelId}/${imageName}/blobs/compressed/${imageTag}.tar.gz`
+  const s3Upload = uploadToExportS3Location(path, gzipStream, logData, metadata)
+
+  // fetch and compress one layer at a time to manage RAM usage
+  // also, gzip can only handle one pipe at a time
+  for (const layer of layers) {
+    const layerDigest = layer['digest']
+    if (!layerDigest || layerDigest.length === 0) {
+      throw InternalError('Could not extract layer digest.', { layer, modelId, imageName, imageTag, ...logData })
+    }
+
+    log.debug('Fetching image layer', {
+      modelId,
+      imageName,
+      imageTag,
+      layerDigest,
+      ...logData,
+    })
+    const responseBody = await getImageBlob(user, modelId, imageName, layerDigest)
+
+    // pipe the body to gzip using streams
+    await new Promise<void>((resolve, reject) => {
+      responseBody.body
+        .on('error', (err) => {
+          reject(InternalError('Error while fetching layer stream', { layerDigest, error: err }))
+        })
+        .on('end', () => {
+          log.debug('Finished fetching layer stream', { layerDigest })
+          // call `resolve()` otherwise the pipe will get stuck
+          resolve()
+        })
+        // pipe to gzip but indicate more data is coming
+        .pipe(gzipStream, { end: false })
+        .on('error', (err) => {
+          reject(InternalError('Error while compressing layer stream', { layerDigest, error: err }))
+        })
+    })
+  }
+  // no more data to write
+  gzipStream.end()
+
+  // wait for the upload to complete
+  await s3Upload
 }
 
 export const ImportKind = {
