+++ backend/src/services/registry.ts
@@ -1,4 +1,4 @@
-import { listImageTags, listModelRepos } from '../clients/registry.js'
+import { getImageTagManifest, getRegistryLayerStream, listImageTags, listModelRepos } from '../clients/registry.js'
 import authorisation from '../connectors/authorisation/index.js'
 import { UserInterface } from '../models/User.js'
 import { getAccessToken } from '../routes/v1/registryAuth.js'
@@ -32,4 +32,46 @@ export async function listModelImages(user: UserInterface, modelId: string) {
   )
 
   return versions
+}
+
+export async function listImageTagLayers(user: UserInterface, modelId: string, imageName: string, imageTag: string) {
+  const model = await getModelById(user, modelId)
+
+  const auth = await authorisation.image(user, model, {
+    type: 'repository',
+    name: modelId,
+    actions: ['pull'],
+  })
+  if (!auth.success) {
+    throw Forbidden(auth.info, { userDn: user.dn, modelId })
+  }
+
+  const repositoryToken = await getAccessToken({ dn: user.dn }, [
+    { type: 'repository', class: '', name: `${modelId}/${imageName}`, actions: ['*'] },
+  ])
+
+  // get which layers exist for the model
+  const manifest = await getImageTagManifest(repositoryToken, { namespace: modelId, image: imageName }, imageTag)
+
+  return manifest.layers
+}
+
+export async function getImageBlob(user: UserInterface, modelId: string, imageName: string, digest: string) {
+  const model = await getModelById(user, modelId)
+
+  const auth = await authorisation.image(user, model, {
+    type: 'repository',
+    name: modelId,
+    actions: ['pull'],
+  })
+  if (!auth.success) {
+    throw Forbidden(auth.info, { userDn: user.dn, modelId })
+  }
+
+  const repositoryToken = await getAccessToken({ dn: user.dn }, [
+    { type: 'repository', class: '', name: `${modelId}/${imageName}`, actions: ['*'] },
+  ])
+  const responseBody = await getRegistryLayerStream(repositoryToken, { namespace: modelId, image: imageName }, digest)
+
+  return responseBody
 }
